refactor: reorganized functions
